Audit panics and add clippy lints (#84)
Audits for additional panics, predominantly with the `embassy` feature
enabled. No breaking changes introduced, in a couple instances I could
convert an unwrap to a `PdError::InvalidParams` which seemed to make the
most sense in those cases.

This PR also additionally addresses a few clippy warnings unrelated to
panics. I had to bump the MSRV from 1.82 to 1.88 since currently there
is already existing code that is not stable under 1.83 which clippy
warned about (and since the code was already introduced, I don't think
there's any harm in bumping the MSRV officially).

Resolves #83